refactoring for stream/await mode
