+++ frontend/src/routes/(app)/(internal)/frameworks/[id=uuid]/report/+page.server.ts
@@ -7,11 +7,15 @@ export const load = (async ({ fetch, params, url }) => {
 	const reportUrl = new URL(`${BASE_API_URL}/frameworks/${params.id}/report/`);
 	if (ig) reportUrl.searchParams.set('implementation_group', ig);
 
-	const res = await fetch(reportUrl.toString());
-	if (!res.ok) {
-		throw error(res.status, `Failed to load framework report (${res.status})`);
-	}
-	const report = await res.json();
+	// Stream the report so the page shell can render immediately with a
+	// spinner; the body fills in once the API responds. For large frameworks
+	// this avoids a multi-second blank screen.
+	const reportPromise = fetch(reportUrl.toString()).then(async (res) => {
+		if (!res.ok) {
+			throw error(res.status, `Failed to load framework report (${res.status})`);
+		}
+		return res.json();
+	});
 
-	return { report };
+	return { stream: { report: reportPromise } };
 }) satisfies PageServerLoad;
